+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## StreamChat
 ### ✅ Added
 - Add a new state layer with async-await and observable state objects ([learn more](https://getstream.io/chat/docs/sdk/ios/state-layer/state-layer-overview/)) [#3177](https://github.com/GetStream/stream-chat-swift/pull/3177)
+### 🐞 Fixed
+- Keep the message order when sending multiple messages and one of them fails [#3202](https://github.com/GetStream/stream-chat-swift/pull/3202)
 
 # [4.55.0](https://github.com/GetStream/stream-chat-swift/releases/tag/4.55.0)
 _May 13, 2024_